Added `ser/lists` module
